Major Refactor
Split out the configuration of the certs so I would not have to
regenerate them every time I wanted to create and destroy the cluster.

One issue is that the ELB DNS name is dynamic and I have to find a
solution for how to include the DNS name in the certs or figure out a
different manner of allowing traffic from the load balancer.